+++ src/lib.rs
@@ -1,0 +1,7 @@
+#![feature(proc_macro_hygiene, decl_macro)]
+#![allow(non_snake_case)]
+#![feature(test)]
+
+pub mod explorers;
+pub mod game;
+pub mod mimc;
