+++ src/mimc.rs
@@ -1,25 +1,6 @@
-#![feature(proc_macro_hygiene, decl_macro)]
-#![feature(test)]
-#![allow(non_snake_case)]
-
-#[macro_use]
-extern crate rocket;
-
-#[macro_use]
-extern crate lazy_static;
-
-// #[macro_use]
-// extern crate log;
 
 use bigint::U512;
-
-use rocket_contrib::json::Json;
-use serde::{Deserialize, Serialize};
-
-use rayon::prelude::*;
-
-use rocket::http::Method;
-use rocket_cors::{catch_all_options_routes, AllowedHeaders, AllowedOrigins};
+use lazy_static::lazy_static;
 
 lazy_static! {
     static ref P: U512 = U512::from_dec_str(
@@ -259,8 +240,8 @@ lazy_static! {
 }
 
 #[derive(Debug, Clone)]
-struct PrimeElem {
-    x: U512,
+pub struct PrimeElem {
+    pub x: U512,
 }
 
 impl PrimeElem {
@@ -291,7 +272,7 @@ impl PrimeElem {
     }
 }
 
-struct MimcState {
+pub struct MimcState {
     l: PrimeElem,
     r: PrimeElem,
     rounds: usize,
@@ -299,6 +280,10 @@ struct MimcState {
 }
 
 impl MimcState {
+    pub fn rarity(planetRarity: u32) -> (U512, bool) {
+        P.overflowing_div(U512::from(planetRarity))
+    }
+
     fn new(rounds: usize, k: PrimeElem) -> MimcState {
         assert!(rounds <= C.len());
         MimcState {
@@ -324,7 +309,7 @@ impl MimcState {
         self.r = t.fifth_power().plus(&self.r);
     }
 
-    fn sponge(inputs: Vec<i64>, n_outputs: usize, rounds: usize) -> Vec<PrimeElem> {
+    pub fn sponge(inputs: Vec<i64>, n_outputs: usize, rounds: usize) -> Vec<PrimeElem> {
         let inputs = inputs
             .into_iter()
             .map(|x| {
@@ -352,548 +337,3 @@ impl MimcState {
         outputs
     }
 }
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-struct Coords {
-    x: i64,
-    y: i64,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-struct Planet {
-    coords: Coords,
-    hash: String,
-}
-
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-struct ChunkFootprint {
-    bottomLeft: Coords,
-    sideLength: i64,
-}
-
-#[allow(non_snake_case)]
-#[derive(Deserialize, Debug)]
-struct Task {
-    chunkFootprint: ChunkFootprint,
-    planetRarity: u32,
-}
-
-#[allow(non_snake_case)]
-#[derive(Serialize)]
-struct Response {
-    chunkFootprint: ChunkFootprint,
-    planetLocations: Vec<Planet>,
-}
-
-#[get("/")]
-fn hello() -> &'static str {
-    "Hello, world!"
-}
-
-#[post("/mine", data = "<task>")]
-fn mine(task: Json<Task>) -> Json<Response> {
-    // info_!("{:?}", task);
-    // Info: Json(Task { chunkFootprint: ChunkFootprint { bottomLeft: Coords { x: -15456, y: 18944 }, sideLength: 32 }, planetRarity: 16384 })
-    // Info: Json(Task { chunkFootprint: ChunkFootprint { bottomLeft: Coords { x: -15456, y: 18912 }, sideLength: 32 }, planetRarity: 16384 })
-    // Info: Json(Task { chunkFootprint: ChunkFootprint { bottomLeft: Coords { x: -15456, y: 18880 }, sideLength: 32 }, planetRarity: 16384 })
-    // Info: Json(Task { chunkFootprint: ChunkFootprint { bottomLeft: Coords { x: -15456, y: 18848 }, sideLength: 32 }, planetRarity: 16384 })
-
-    let x = task.chunkFootprint.bottomLeft.x;
-    let y = task.chunkFootprint.bottomLeft.y;
-    let size = task.chunkFootprint.sideLength;
-
-    let (threshold, overflowed) = P.overflowing_div(U512::from(task.planetRarity));
-    assert!(!overflowed);
-
-    let planets = (x..(x + size))
-        .into_par_iter()
-        .map(|xi| {
-            let mut planets = Vec::new();
-            for yi in y..(y + size) {
-                let hash = MimcState::sponge(vec![xi, yi], 1, 220)[0].x;
-                if hash < threshold {
-                    planets.push(Planet {
-                        coords: Coords { x: xi, y: yi },
-                        hash: hash.to_string(),
-                    });
-                }
-            }
-            planets
-        })
-        .flatten()
-        .collect::<Vec<_>>();
-
-    Json(Response {
-        chunkFootprint: task.chunkFootprint.clone(),
-        planetLocations: planets,
-    })
-}
-
-fn main() {
-    // for x in 0.. {
-    //     if x % 100 == 0 {
-    //         println!("trying ({}, 0)", x);
-    //     }
-    //     MimcState::sponge(vec![x, 0], 1, 220);
-    // }
-
-    // println!("{:?}", MimcState::sponge(vec![-2048, 0], 1, 220));
-    let allowed_origins = AllowedOrigins::all();
-    let cors = rocket_cors::CorsOptions {
-        allowed_origins,
-        allowed_methods: vec![Method::Post].into_iter().map(From::from).collect(),
-        allowed_headers: AllowedHeaders::all(),
-        allow_credentials: true,
-        ..Default::default()
-    }
-    .to_cors()
-    .unwrap();
-    let options_routes = catch_all_options_routes();
-
-    rocket::ignite()
-        .mount("/", routes![mine, hello])
-        .mount("/", options_routes)
-        .manage(cors.clone())
-        .attach(cors)
-        .launch();
-}
-
-trait MiningPattern {
-    fn next(&self, fromChunk: &ChunkFootprint) -> ChunkFootprint;
-}
-
-struct SpiralMiner {
-    currentChunk: ChunkFootprint,
-    pattern: Spiral,
-}
-
-#[allow(unused)]
-impl SpiralMiner {
-    fn new(center: Coords, chunkSideLength: u16) -> Self {
-        let currentChunk = ChunkFootprint {
-            bottomLeft: center.clone(),
-            sideLength: chunkSideLength as i64,
-        };
-        let pattern = Spiral::new(&center, chunkSideLength);
-
-        Self {
-            currentChunk,
-            pattern,
-        }
-    }
-}
-impl Iterator for SpiralMiner {
-    type Item = ChunkFootprint;
-    fn next(&mut self) -> Option<Self::Item> {
-        let next = self.pattern.next(&self.currentChunk);
-        self.currentChunk = next.clone();
-        Some(next)
-    }
-}
-
-struct Spiral {
-    chunkSideLength: u16,
-    fromChunk: ChunkFootprint,
-}
-
-#[allow(unused)]
-impl Spiral {
-    fn new(center: &Coords, chunkSideLength: u16) -> Self {
-        //floor by default?
-
-        let length = i64::from(chunkSideLength);
-
-        let bottomLeftX = (center.x / length) * length;
-        let bottomLeftY = (center.y / length) * length;
-        let bottomLeft = Coords {
-            x: bottomLeftX,
-            y: bottomLeftY,
-        };
-
-        let fromChunk = ChunkFootprint {
-            bottomLeft,
-            sideLength: length,
-        };
-
-        Self {
-            fromChunk,
-            chunkSideLength,
-        }
-    }
-}
-impl MiningPattern for Spiral {
-    fn next(&self, chunk: &ChunkFootprint) -> ChunkFootprint {
-        let homeX = self.fromChunk.bottomLeft.x;
-        let homeY = self.fromChunk.bottomLeft.y;
-        let currX = chunk.bottomLeft.x;
-        let currY = chunk.bottomLeft.y;
-
-        let mut nextBottomLeft = Coords { x: currX, y: currY };
-
-        let length = i64::from(self.chunkSideLength);
-
-        if currX == homeX && currY == homeY {
-            nextBottomLeft.y = homeY + length;
-        } else if currY - currX > homeY - homeX && currY + currX >= homeX + homeY {
-            if currY + currX == homeX + homeY {
-                // break the circle
-                nextBottomLeft.y = currY + length;
-            } else {
-                nextBottomLeft.x = currX + length;
-            }
-        } else if currX + currY > homeX + homeY && currY - currX <= homeY - homeX {
-            nextBottomLeft.y = currY - length;
-        } else if currX + currY <= homeX + homeY && currY - currX < homeY - homeX {
-            nextBottomLeft.x = currX - length;
-        } else {
-            // if (currX + currY < homeX + homeY && currY - currX >= homeY - homeX)
-            nextBottomLeft.y = currY + length;
-        }
-
-        ChunkFootprint {
-            bottomLeft: nextBottomLeft,
-            sideLength: length,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn sixteen_iter() {
-        let center = Coords { x: 0, y: 0 };
-        let chunkSideLength = 16;
-        let mut miner = SpiralMiner::new(center, chunkSideLength);
-
-        assert_eq!(
-            miner.next(),
-            Some(ChunkFootprint {
-                bottomLeft: Coords {
-                    x: 0,
-                    y: chunkSideLength as i64,
-                },
-                sideLength: chunkSideLength as i64
-            })
-        );
-
-        assert_eq!(
-            miner.next(),
-            Some(ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: chunkSideLength as i64,
-                },
-                sideLength: chunkSideLength as i64
-            })
-        );
-
-        assert_eq!(
-            miner.next(),
-            Some(ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: 0,
-                },
-                sideLength: chunkSideLength as i64
-            })
-        );
-
-        assert_eq!(
-            miner.next(),
-            Some(ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: -(chunkSideLength as i64),
-                },
-                sideLength: chunkSideLength as i64
-            })
-        );
-
-        assert_eq!(
-            miner.next(),
-            Some(ChunkFootprint {
-                bottomLeft: Coords {
-                    x: 0,
-                    y: -(chunkSideLength as i64),
-                },
-                sideLength: chunkSideLength as i64
-            })
-        );
-    }
-
-    #[test]
-    fn sixteen() {
-        let chunkSideLength = 16;
-        let center = Coords { x: 0, y: 0 };
-
-        let start = ChunkFootprint {
-            bottomLeft: center.clone(),
-            sideLength: chunkSideLength as i64,
-        };
-        let spiral = Spiral::new(&center, chunkSideLength);
-        let first = spiral.next(&start);
-        let second = spiral.next(&first);
-        let third = spiral.next(&second);
-        let fourth = spiral.next(&third);
-        let fifth = spiral.next(&fourth);
-
-        assert_eq!(
-            first,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: 0,
-                    y: chunkSideLength as i64,
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            second,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: chunkSideLength as i64,
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            third,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: 0,
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            fourth,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: -(chunkSideLength as i64),
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            fifth,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: 0,
-                    y: -(chunkSideLength as i64),
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-    }
-
-    #[test]
-    fn thirtytwo() {
-        let chunkSideLength = 32;
-        let center = Coords { x: 0, y: 0 };
-
-        let start = ChunkFootprint {
-            bottomLeft: center.clone(),
-            sideLength: chunkSideLength as i64,
-        };
-        let spiral = Spiral::new(&center, chunkSideLength);
-        let first = spiral.next(&start);
-        let second = spiral.next(&first);
-        let third = spiral.next(&second);
-        let fourth = spiral.next(&third);
-        let fifth = spiral.next(&fourth);
-
-        assert_eq!(
-            first,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: 0,
-                    y: chunkSideLength as i64,
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            second,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: chunkSideLength as i64,
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            third,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: 0,
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            fourth,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: chunkSideLength as i64,
-                    y: -(chunkSideLength as i64),
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-
-        assert_eq!(
-            fifth,
-            ChunkFootprint {
-                bottomLeft: Coords {
-                    x: 0,
-                    y: -(chunkSideLength as i64),
-                },
-                sideLength: chunkSideLength as i64
-            }
-        );
-    }
-
-    extern crate test;
-    use test::Bencher;
-    #[bench]
-    fn sixteen_bench(b: &mut Bencher) {
-        let center = Coords { x: 0, y: 0 };
-        let chunkSideLength = 16;
-        let mut miner = SpiralMiner::new(center, chunkSideLength);
-
-        b.iter(|| {
-            let n = test::black_box(100000);
-
-            for i in 0..n {
-                let _ = miner.next().unwrap();
-            }
-        })
-    }
-
-    #[bench]
-    fn thirtytwo_bench(b: &mut Bencher) {
-        let center = Coords { x: 0, y: 0 };
-        let chunkSideLength = 32;
-        let mut miner = SpiralMiner::new(center, chunkSideLength);
-
-        b.iter(|| {
-            let n = test::black_box(100000);
-
-            for i in 0..n {
-                let _ = miner.next().unwrap();
-            }
-        })
-    }
-
-    #[bench]
-    fn sixtyfour_bench(b: &mut Bencher) {
-        let center = Coords { x: 0, y: 0 };
-        let chunkSideLength = 64;
-        let mut miner = SpiralMiner::new(center, chunkSideLength);
-
-        b.iter(|| {
-            let n = test::black_box(100000);
-
-            for i in 0..n {
-                let _ = miner.next().unwrap();
-            }
-        })
-    }
-
-    #[bench]
-    fn onetwentyeigth_bench(b: &mut Bencher) {
-        let center = Coords { x: 0, y: 0 };
-        let chunkSideLength = 128;
-        let mut miner = SpiralMiner::new(center, chunkSideLength);
-
-        b.iter(|| {
-            let n = test::black_box(100000);
-
-            for i in 0..n {
-                let _ = miner.next().unwrap();
-            }
-        })
-    }
-
-    #[bench]
-    fn twofiftysix_bench(b: &mut Bencher) {
-        let center = Coords { x: 0, y: 0 };
-        let chunkSideLength = 256;
-        let mut miner = SpiralMiner::new(center, chunkSideLength);
-
-        b.iter(|| {
-            let n = test::black_box(100000);
-
-            for i in 0..n {
-                let _ = miner.next().unwrap();
-            }
-        })
-    }
-}
-
-// private async nextValidExploreTarget(
-//     chunkLocation: ChunkFootprint,
-//     jobId: number
-//   ): Promise<ChunkFootprint | null> {
-//     // returns the first valid chunk equal to or after `chunk` (in the explore order of mining pattern) that hasn't been explored
-//     // async because it may take indefinitely long to find the next target. this will block UI if done sync
-//     // we use this trick to promisify:
-//     // https://stackoverflow.com/questions/10344498/best-way-to-iterate-over-an-array-without-blocking-the-ui/10344560#10344560
-
-//     // this function may return null if user chooses to stop exploring or changes mining pattern in the middle of its resolution
-//     // so any function calling it should handle the null case appropriately
-//     let candidateChunk = chunkLocation;
-//     let count = 10000;
-//     while (!this.isValidExploreTarget(candidateChunk) && count > 0) {
-//       candidateChunk = this.miningPattern.nextChunk(candidateChunk);
-//       count -= 1;
-//     }
-//     // since user might have switched jobs or stopped exploring during the above loop
-//     if (!this.isExploring && jobId !== this.currentJobId) {
-//       return null;
-//     }
-//     if (this.isValidExploreTarget(candidateChunk)) {
-//       return candidateChunk;
-//     }
-//     return new Promise((resolve) => {
-//       setTimeout(async () => {
-//         const nextNextChunk = await this.nextValidExploreTarget(
-//           candidateChunk,
-//           jobId
-//         );
-//         resolve(nextNextChunk);
-//       }, 0);
-//     });
-//   }
-
-//   private exploreNext(fromChunk: ChunkFootprint, jobId: number) {
-//     this.nextValidExploreTarget(fromChunk, jobId).then(
-//       (nextChunk: ChunkFootprint | null) => {
-//         if (!!nextChunk) {
-//           const nextChunkKey = this.chunkLocationToKey(nextChunk, jobId);
-//           this.exploringChunk[nextChunkKey] = {
-//             chunkFootprint: nextChunk,
-//             planetLocations: [],
-//           };
-//           this.exploringChunkStart[nextChunkKey] = Date.now();
-//           this.minersComplete[nextChunkKey] = 0;
-//           this.sendMessageToWorkers(nextChunk, jobId);
-//         }
-//       }
-//     );
-//   }
